telegram-bot-mongodb/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pymongo==4.5.0
dnspython==2.4.2
flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
apscheduler==3.10.4
